little change in the API:
In order to develop a way for the user to keep a state machine or manage
connection sessions this patch introduce a new structure: Connection.

This structure helps us to manage Connection without any knowledge of
the type of the Connection (TCP/UDP) from the Server point of view.
It also embeds some information data as: creation date, last used date.
And provide 2 tools:
 * setKeepOpen: in order to say to the server if we want to keep this
   connection open or not
 * getContext: a "MVar a" which can be use by the user to store/read any
   state information.